2023/03/24
Additions and changes to look out for this week:
•	Communications and Event management - rearrange
•	Remove glossary links for self-explanatory words
•	Glossary Approaches / Leads - see tracking key milestones under LEADS (this is the only glossary link)
•	Add features to communications – to be reviewed
•	Add Event management overview
•	Add detail to SARS tax details and issue tax certificate
o	Under New field requirements added - The following are the new fields that SARS requires to be added to tax certificates
o	Before Issue a tax certificate Xplain added - NOTE: If all details required by SARS are not completed, the tax certificate will not be generated.
•	Workflow:
o	Changed links for glossary and workflow
o	Under add a workflow added: Please contact our support desk to review your workflow or for assistance to set it up.
o	Began work on marking old content as obsolete so it doesn't appear in the Help Search